requirements.txt
flask
numpy
opencv-python
pandas
Pillow
pyaadhaar
pytesseract
pyzbar
requests
torch
ultralytics
